requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
chillin-client==1.3.1
chillin-server==2.2.0
circuits==3.2
click==7.1.2
colorama==0.4.3
configparser==3.5.0
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-filter==2.4.0
Django-Verify-Email==0.0.5
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
enum34==1.1.6
filelock==3.0.12
gevent==20.9.0
greenlet==0.4.17
h11==0.11.0
haikunator==2.1.0
hazm==0.7.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.4.2
joblib==0.17.0
koala-serializer==0.8.0
lazy-object-proxy==1.4.3
Markdown==3.3.2
mccabe==0.6.1
msgpack==1.0.0
nltk==3.4
Pillow==8.0.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDbLite==3.0.4
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyodbc==4.0.30
pyOpenSSL==20.0.0
pypiwin32==223
pytz==2020.1
pywin32==300
redis==3.5.3
regex==2020.11.13
requests==2.25.0
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
tqdm==4.54.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
uvicorn==0.13.0
virtualenv==20.2.1
websocket==0.2.1
websocket-client==0.57.0
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.2.0
